Scale image down if >= 1000 pixels wide.
The character to be used is decided by getting the average brightness
in a 2x2 square of pixels. The size used now is equal to the width of
the image divided by 500. It defaults to 2x2 if the image is less than
500 pixels wide.

Previously, larger images took too long to process, this change provides
a maximum time an image should take due to the scaling.